requirements.txt
torch>=0.4.1
torchvision>=0.2.0
Pillow>=5.0.0
numpy>=1.8
dominate>=2.3.1
visdom>=0.1.7
python-dateutil >= 1.0.0